Added interfaces to client.go
